Add dialog component, implement rating recipes & adding recipes to lists

+++ server/src/modules/recipe/index.js
@@ -83,7 +83,7 @@ const typeDefs = gql`
 
   type Rating {
     user: User!
-    score: Int!
+    score: Float!
   }
 
   enum Difficulty {
@@ -151,7 +151,7 @@ const typeDefs = gql`
 
   input RecipeRatingInput {
     id: ID!
-    score: Int!
+    score: Float!
   }
 `;
